+++ src/util/parseAttributeMap.ts
@@ -46,11 +46,36 @@ export function parseSamlAttributeMap(
 export function parseAuthenticationMethod(
   authenticationMethod: string
 ): number {
-  if (authenticationMethod == "Ticket") return 0;
-  if (authenticationMethod == "static") return 1;
-  if (authenticationMethod == "dynamic") return 2;
-  if (authenticationMethod == "SAML") return 3;
-  if (authenticationMethod == "JWT") return 4;
+  if (authenticationMethod.toLowerCase() == "ticket") return 0;
+  if (authenticationMethod.toLowerCase() == "static") return 1;
+  if (authenticationMethod.toLowerCase() == "headerstaticuserdirectory")
+    return 1;
+  if (authenticationMethod.toLowerCase() == "dynamic") return 2;
+  if (authenticationMethod.toLowerCase() == "headerdynamicuserdirectory")
+    return 2;
+  if (authenticationMethod.toLowerCase() == "saml") return 3;
+  if (authenticationMethod.toLowerCase() == "jwt") return 4;
+
+  throw new Error(
+    `virtualProxy.create: "authenticationMethod" not found "${authenticationMethod}"`
+  );
+}
+
+export function parseSameSiteAttribute(sameSiteAttribute: string): number {
+  if (sameSiteAttribute.toLowerCase() == "no attribute") return 0;
+  if (sameSiteAttribute.toLowerCase() == "none") return 1;
+  if (sameSiteAttribute.toLowerCase() == "lax") return 2;
+  if (sameSiteAttribute.toLowerCase() == "strict") return 3;
 
   return 0;
+}
+
+export function parseAnonymousAccessMode(anonymousAccessMode: string): number {
+  if (anonymousAccessMode.toLowerCase() == "no anonymous") return 0;
+  if (anonymousAccessMode.toLowerCase() == "allow anonymous") return 1;
+  if (anonymousAccessMode.toLowerCase() == "always anonymous") return 2;
+
+  throw new Error(
+    `virtualProxy.create: "anonymousAccessMode" not found "${anonymousAccessMode}"`
+  );
 }